Planes cant take off/land in bad weather
